requirements.txt
-e git+https://github.com/django/django.git#egg=Django
django-celery>=3.0.11
ipython>=0.13.1
lxml>=3.1beta1
nltk>=2.0.4
numpy>=1.6.2
pdfminer>=20110515
readability-lxml>=0.2.6.1
requests>=1.1.0
scikit-learn>=0.13
scipy>=0.11.0
slate>=0.3
